Refactor file structure for EmbedderView and related files
