+++ src/views/EmbedderView/components/EmbedController.js
@@ -4,8 +4,8 @@ import {
   Typography, Paper, withStyles, InputBase, Divider, Button,
 } from '@material-ui/core';
 import { FormattedMessage } from 'react-intl';
-import SMRadio from '../../components/SMRadio';
-import styles from './styles';
+import SMRadio from '../../../components/SMRadio';
+import styles from '../styles';
 
 
 const customStyles = () => ({
